refactor(storage): add readSource field accessors
